Fixed lab1 - all ints to float - try blocks in main

+++ c++/lab1/queue1.hpp
@@ -6,8 +6,8 @@ class Queue1 : public Queue {
 public:
     Queue1();
 
-    double average();
-    void   printAverage();
+    float average();
+    void  printAverage();
 
 };
